V[2.0]冲刺的增删改查 分页查询基础能力

+++ eladmin-system/src/main/java/me/zhengjie/modules/system/dao/mapper/BusinessLineMapper.java
@@ -11,9 +11,9 @@ public interface BusinessLineMapper {
 
     Integer insertBusinessLine(BusinessLineDO businessLineDO);
 
-    BusinessLineDO findByBusinessLineId(Long id);
+    BusinessLineDO findByBusinessLineId(int id);
 
-    Integer deleteBusinessLine(Long id);
+    Integer deleteBusinessLine(int id);
 
 //    BusinessLineDO findBusinessLineById(@Param("id") int id);
 //
